requirements.txt
dominate==2.7.0
numpy==1.19.5
opencv_python==4.3.0.38
opencv-contrib-python==4.3.0.38
Pillow==11.2.1
scipy==1.5.4
torch==1.10.1
torchvision==0.11.2
visdom==0.2.4
